+++ sdk/core/azure-core/inc/azure/core/context.hpp
@@ -392,11 +392,17 @@ namespace Azure { namespace Core {
     }
 
     /**
+     * @brief Check if the context is canceled.
+     * @return `true` if this context is canceled, `false` otherwise.
+     */
+    bool IsCanceled() const { return CancelWhen() < std::chrono::system_clock::now(); }
+
+    /**
      * @brief Throw an exception if the context was canceled.
      */
     void ThrowIfCanceled() const
     {
-      if (CancelWhen() < std::chrono::system_clock::now())
+      if (IsCanceled())
       {
         throw OperationCanceledException("Request was canceled by context.");
       }