+++ sdk/core/azure-core/inc/azure/core/credentials.hpp
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include <azure/core/context.hpp>
+#include "azure/core/context.hpp"
 
 #include <chrono>
 #include <memory>