+++ sdk/core/azure-core/test/ut/string.cpp
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/core/strings.hpp"
-#include "gtest/gtest.h"
+#include <azure/core/strings.hpp>
+#include <gtest/gtest.h>
 #include <string>
 
 TEST(String, invariantCompare)