+++ sdk/core/azure-core/test/ut/transport_adapter_base.hpp
@@ -9,9 +9,8 @@
  *
  */
 
-#include "gtest/gtest.h"
+#include <gtest/gtest.h>
 #include <azure/core/http/body_stream.hpp>
-#include <azure/core/http/curl/curl.hpp>
 #include <azure/core/http/http.hpp>
 #include <azure/core/http/pipeline.hpp>