+++ sdk/core/azure-core/CHANGELOG.md
@@ -4,12 +4,15 @@
 
 ### New Features
 
+- Added a WinHTTP-based `HttpTransport` called `WinHttpTransport` and use that as the default `TransportPolicyOptions.Transport` on Windows when sending and receiving requests and responses over the wire.
 - Added support for long running operations with `Operation<T>`
 
 ### Breaking Changes
 
 - Removed `DateTime::operator Duration()`.
 - Moved `Azure::Core::BearerTokenAuthenticationPolicy`, defined in `azure/core/credentials.hpp` to `Azure::Core::Http` namespace in `azure/core/http/policy.hpp` header.
+- Removed option `AllowBeast` from `CurlTransportSSLOptions` in `CurlTransportOptions`.
+- Changed default option `NoRevoke` from `CurlTransportSSLOptions` for the `CurlTransportOptions` to `true`. This disables the revocation list checking by default.
 
 ## 1.0.0-beta.3 (2020-11-11)